more restructuring of files
